Move different package constructions out of utils
