+++ src/context.js
@@ -15,4 +15,9 @@ module.exports = {
   ssoPaginate: require("./models/ssoPaginate.js"),
   ssoRedirect: require("./models/ssoRedirect.js"),
   ssoHTML: require("./models/ssoHTML.js"),
+  models: {
+    constructPackageObjectFull: require("./models/constructPackageObjectFull.js"),
+    constructPackageObjectShort: require("./models/constructPackageObjectShort.js"),
+    constructPackageObjectJSON: require("./models/constructPackageObjectJSON.js"),
+  }
 };
